refactor(datepicker): name refactoring
