+++ components/datepicker/index.js
@@ -1,2 +1,2 @@
-export { default as DtDatepicker } from './datepicker.vue';
+export { default as DtDatepicker } from './DtDatepicker.vue';
 export {} from './datepicker_constants';
